Allow for 720p as well as a show year versions in the file name string
Limiting the initial name match means the season/episode section doesn't
pick up 720 as that match instead. The literal period after the
season/episode section along with making the episode identifier required
allows for both 720 and a year version in the same filename.
Unfortunately this removes the ability to work with files in the format
series.000 for now, but that is far more of an edge case.