Return a more sensible output when the config's defaults are missing
This is a stop gap for now as no defaults shouldn't really cause the
program to exit, but ideally would have some inbuilt defaults to cope
with a situation like this. Either in the code or in a config included
with the code to make it more modular. I'll expand on this in #10.

Fixes #9