fallback default filename format
The documentation suggests that a default format is used if none is specified
in the config file, but in fact "format" in "build_path" just becomes False if
none is specified. This hard-codes the documented default format -- I'm not sure
if there's a better system in your config system for specifying these defaults.