Pass around details taken from the filename as a dict
Expand the filename details dict when creating an episode and pick them
up as kwargs in the __init__(). If any of the core variables aren't
don't exist in the kwargs then don't set them or we'll break the
__getattr__ since the variable will exist on the episode object.
If show_name, season or episode aren't set on an Episode add a decent
error message.

Expand the result of Main.extract_details_from_file when creating an
episode object.
Catch AttributeErrors in the except that doesn't exit the program but
print the error if there are args in the Exception. This means we can
throw an AttributeError in Episode.__getattr__ but create logs.

Don't assume each group will exist in the regex so build up the filename
credentials dict incrementally. Allows for nicer logs too and a cleaner
exit when there are no matches.