Fix partial matching regression bug
When multiple episode matching was added it introduced the File object
that forced the 4 parts of a filename to be required arguments when
creating a File. This broke the --partial options.

These changes also remove any trailing digits from the episode number
from within the codebase to avoid confusion. This is much easier now
that the output formatting has been fixed.