tvdb may return multiple series
In cases where TheTVDB.com returns multiple entries *and* the series we
are looking for is not the first one, the series would not be found,
because only the first one would be checked.

In my case, I am looking for 'The Good Place' and TheTVDB.com returns
the following structure (in XML equivalent).

    - Series
      - SeriesName: The Good Place: The Podcast
    - Series
      - SeriesName: The Good Place

I am looking for the second one, but the loop stops at the first entry
when raising the ShowNotFoundException.

Solve this problem by raising the exception only when all series are
checked.